mvp/original/SQLReleaseNotes2/requirements.txt
pyodbc>=4.0.32
openai>=1.2.0
python-dotenv>=1.0.0
tqdm>=4.65.0
demjson3>=3.0.5
argparse>=1.4.0
colorama>=0.4.6
